requirements.txt
python-nmap
pyOpenSSL
python-snap7
pymodbus